+++ CHANGELOG.rst
@@ -1,5 +1,10 @@
 Change Log
 ----------
+**0.20** (2020-05-12)
+
+- All objects can now print their values
+- Add parser + object for generic profile
+
 **0.19** (2020-05-03)
 
 - Add following missing elements to telegram specification v4:
@@ -45,7 +50,7 @@ Change Log
 
 **0.10** (2017-06-05)
 
-- bugix: don't force full telegram signatures (`pull request #25 <https://github.com/ndokter/dsmr_parser/pull/25>`_)
+- bugfix: don't force full telegram signatures (`pull request #25 <https://github.com/ndokter/dsmr_parser/pull/25>`_)
 - removed unused code for automatic telegram detection as this needs reworking after the fix mentioned above
 - InvalidChecksumError's are logged as warning instead of error